Implement delete template method
